requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
drf-yasg
django-filter
sqlparse==0.5.4
tzdata==2025.2
requests==2.32.3
requests>=2.31.0

